+++ examples/02_typescript/src/App.tsx
@@ -1,19 +1,18 @@
 import React, { Suspense } from 'react';
-
-import { ErrorBoundary } from 'react-hooks-fetch';
+import { ErrorBoundary, FallbackProps } from 'react-error-boundary';
 
 import Item, { ErrorItem } from './Item';
 
-const renderError = (error: Error, retry: () => void) => (
+const ErrorFallback = ({ error, resetErrorBoundary }: FallbackProps) => (
   <div>
     <h1>Something went wrong.</h1>
     <p>{error.message}</p>
-    <button type="button" onClick={retry}>Retry</button>
+    <button type="button" onClick={resetErrorBoundary}>Retry</button>
   </div>
 );
 
 const App: React.FC = () => (
-  <ErrorBoundary fallback={renderError}>
+  <ErrorBoundary FallbackComponent={ErrorFallback}>
     <Suspense fallback={<span>Loading...</span>}>
       <Item />
       <hr />
@@ -23,7 +22,7 @@ const App: React.FC = () => (
       <hr />
       <ErrorItem />
       <hr />
-      <ErrorBoundary fallback={renderError}>
+      <ErrorBoundary FallbackComponent={ErrorFallback}>
         <ErrorItem />
       </ErrorBoundary>
     </Suspense>