+++ examples/02_typescript/src/DisplayData.tsx
@@ -1,5 +1,5 @@
 // eslint-disable-next-line spaced-comment
-/// <reference types="react/experimental" />
+/// <reference types="react/next" />
 
 import React, { useTransition } from 'react';
 
@@ -14,9 +14,7 @@ type Props = {
 };
 
 const DisplayData: React.FC<Props> = ({ id, result, refetch }) => {
-  const [startTransition, isPending] = useTransition({
-    timeoutMs: 1000,
-  });
+  const [isPending, startTransition] = useTransition();
   const onClick = () => {
     startTransition(() => {
       refetch(id);