+++ src/index.ts
@@ -1,3 +1,2 @@
-export { ErrorBoundary } from './ErrorBoundary';
 export { createFetchStore } from './createFetchStore';
 export { useFetch } from './useFetch';